requirements.txt
asgiref==3.8.1
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
Django==3.2.23
djangorestframework==3.12.4
folium==0.19.4
fonttools==4.55.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
numpy==2.0.2
openrouteservice==2.3.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xyzservices==2024.9.0
zipp==3.21.0
